requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
pydantic-settings==2.3.4
google-generativeai==0.7.2
httpx==0.27.0
python-dotenv==1.0.1
